+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/component/RobotConsoleWindow.java
@@ -2,6 +2,7 @@ package de.honoka.qqrobot.starter.component;
 
 import de.honoka.qqrobot.framework.Framework;
 import de.honoka.qqrobot.starter.RobotBasicProperties;
+import de.honoka.sdk.util.code.ThrowsRunnable;
 import de.honoka.sdk.util.system.gui.ConsoleWindow;
 import de.honoka.sdk.util.text.TextUtils;
 import lombok.AccessLevel;
@@ -20,9 +21,13 @@ import java.util.function.Consumer;
 @Accessors(chain = true)
 public class RobotConsoleWindow {
 
-    @Getter
+    @Getter(AccessLevel.NONE)
     @Setter(AccessLevel.NONE)
-    private static ConsoleWindow console;
+    private static ConsoleWindow consoleWindow;
+
+    @Getter(AccessLevel.NONE)
+    @Setter(AccessLevel.NONE)
+    private static RobotConsole console;
 
     private String name;
 
@@ -36,47 +41,82 @@ public class RobotConsoleWindow {
 
     private Consumer<ApplicationContext> onExit = context -> {};
 
+    private ThrowsRunnable beforeRunApplication = () -> {};
+
     private boolean showOnCreate = true;
 
     @Setter(AccessLevel.NONE)
     private ApplicationContext context;
 
-    public RobotConsoleWindow(String name, Class<?> springBootMainClass) {
-        this.name = name;
-        this.springBootMainClass = springBootMainClass;
+    /**
+     * 强制不使用GUI窗口
+     */
+    private boolean forceNoGui = false;
+
+    private RobotConsoleWindow() {}
+
+    public static RobotConsoleWindow of(String name, Class<?> springBootMainClass) {
+        RobotConsoleWindow window = new RobotConsoleWindow();
+        window.name = name;
+        window.springBootMainClass = springBootMainClass;
+        return window;
     }
 
-    public RobotConsoleWindow(String name, double screenZoomScale,
-                              Class<?> springBootMainClass) {
-        this.name = name;
-        this.screenZoomScale = screenZoomScale;
-        this.springBootMainClass = springBootMainClass;
+    public static RobotConsoleWindow of(String name, double screenZoomScale,
+                                        Class<?> springBootMainClass) {
+        RobotConsoleWindow window = new RobotConsoleWindow();
+        window.name = name;
+        window.screenZoomScale = screenZoomScale;
+        window.springBootMainClass = springBootMainClass;
+        return window;
+    }
+
+    private void startApplication() {
+        beforeRunApplication.run();
+        SpringApplication app = new SpringApplication(springBootMainClass);
+        context = app.run(applicationArgs);
     }
 
     public void create() {
+        if(forceNoGui) {
+            console = new RobotConsole();
+            startApplication();
+            return;
+        }
         //创建并显示窗口
-        console = new ConsoleWindow(name, iconPath, () -> {
-            onExit.accept(context);
-        });
-        console.setAutoScroll(true);
-        console.setScreenZoomScale(screenZoomScale);
-        if(showOnCreate) {
-            console.show();
+        try {
+            consoleWindow = ConsoleWindow.Builder.of(name).setOnExit(() -> {
+                onExit.accept(context);
+            }).setScreenZoomScale(screenZoomScale).build();
+            consoleWindow.setAutoScroll(true);
+        } catch(Throwable t) {
+            //若不能加载窗口，则直接以控制台方式运行
+            console = new RobotConsole();
+            startApplication();
+            return;
         }
         //启动应用
-        SpringApplication app = new SpringApplication(springBootMainClass);
-        context = app.run(applicationArgs);
+        startApplication();
         //添加托盘图标菜单项
         Framework<?> framework = context.getBean(Framework.class);
         RobotBasicProperties basicProperties = context.getBean(
                 RobotBasicProperties.class);
-        console.addTrayIconMenuItem("重新登录", true,
+        consoleWindow.addTrayIconMenuItem("重新登录", true,
                 framework::reboot);
-        console.addTrayIconMenuItem("发送测试消息",
+        consoleWindow.addTrayIconMenuItem("发送测试消息",
                 false, () -> {
             String time = TextUtils.getSimpleDateFormat().format(new Date());
             framework.sendGroupMsg(basicProperties.getDevelopingGroup(),
                     time + "\n测试消息");
         });
+    }
+
+    public static String getConsoleText() {
+        if(consoleWindow != null) {
+            return consoleWindow.getText();
+        } else if(console != null) {
+            return console.getText();
+        }
+        return null;
     }
 }